Move all rotational information to poliastro.frames
Dropped two functions in poliastro.coordinates
and corresponding tests.
