+++ src/poliastro/bodies.py
@@ -115,42 +115,10 @@ class Body(_Body):
 
 
 class SolarSystemBody(Body):
-    def rot_elements_at_epoch(self, epoch):
-        """Provides rotational elements at epoch.
-
-        Provides north pole of body and angle to prime meridian.
-
-        Parameters
-        ----------
-        epoch : ~astropy.time.Time, optional
-            Epoch, default to J2000.
-
-        Returns
-        -------
-        ra, dec, W: tuple (~astropy.units.Quantity)
-            Right ascension and declination of north pole, and angle of the prime meridian.
-
-        """
-        T = (epoch.tdb - constants.J2000).to("day").value / 36525
-        d = (epoch.tdb - constants.J2000).to("day").value
-        return self._rot_elements_at_epoch(T, d)
-
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        raise NotImplementedError("Function only defined for some Solar System bodies")
+    pass
 
 
-class _Sun(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        ra = 286.13 * u.deg
-        dec = 63.87 * u.deg
-        W = (84.176 + 14.1844000 * d) * u.deg
-
-        return ra, dec, W
-
-
-Sun = _Sun(
+Sun = SolarSystemBody(
     parent=None,
     k=constants.GM_sun,
     name="Sun",
@@ -162,178 +130,7 @@ Sun = _Sun(
 )
 
 
-class _Mercury(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        M1 = (174.7910857 + 4.092335 * d) * u.deg
-        M2 = (349.5821714 + 8.184670 * d) * u.deg
-        M3 = (164.3732571 + 12.277005 * d) * u.deg
-        M4 = (339.1643429 + 16.369340 * d) * u.deg
-        M5 = (153.9554286 + 20.461675 * d) * u.deg
-        ra = (281.0103 - 0.0328 * T) * u.deg
-        dec = (61.45 - 0.005 * T) * u.deg
-        W = (329.5988 + 6.1385108 * d) * u.deg + (
-            0.01067257 * math.sin(M1.to("rad").value)
-            - 0.00112309 * math.sin(M2.to("rad").value)
-            - 0.00011040 * math.sin(M3.to("rad").value)
-            - 0.00002539 * math.sin(M4.to("rad").value)
-            - 0.00000571 * math.sin(M5.to("rad").value)
-        ) * u.deg
-
-        return ra, dec, W
-
-
-class _Venus(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        ra = 272.76 * u.deg
-        dec = 67.16 * u.deg
-        W = (160.20 - 1.4813688 * d) * u.deg
-
-        return ra, dec, W
-
-
-class _Earth(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        ra = (0.00 - 0.641 * T) * u.deg
-        dec = (90.00 - 0.557 * T) * u.deg
-        W = (190.147 + 360.9856235 * d) * u.deg
-
-        return ra, dec, W
-
-
-class _Mars(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        M1 = (198.991226 + 19139.4819985 * T) * u.deg
-        M2 = (226.292679 + 38280.8511281 * T) * u.deg
-        M3 = (249.663391 + 57420.7251593 * T) * u.deg
-        M4 = (266.183510 + 76560.6367950 * T) * u.deg
-        M5 = (79.398797 + 0.5042615 * T) * u.deg
-
-        ra = (
-            317.269202
-            - 0.10927547 * T
-            + 0.000068 * math.sin(M1.to("rad").value)
-            + 0.000238 * math.sin(M2.to("rad").value)
-            + 0.000052 * math.sin(M3.to("rad").value)
-            + 0.000009 * math.sin(M4.to("rad").value)
-            + 0.419057 * math.sin(M5.to("rad").value)
-        ) * u.deg
-
-        K1 = (122.433576 + 19139.9407476 * T) * u.deg
-        K2 = (43.058401 + 38280.8753272 * T) * u.deg
-        K3 = (57.663379 + 57420.7517205 * T) * u.deg
-        K4 = (79.476401 + 76560.6495004 * T) * u.deg
-        K5 = (166.325722 + 0.5042615 * T) * u.deg
-
-        dec = (
-            54.432516
-            - 0.05827105 * T
-            + 0.000051 * math.cos(K1.to("rad").value)
-            + 0.000141 * math.cos(K2.to("rad").value)
-            + 0.000031 * math.cos(K3.to("rad").value)
-            + 0.000005 * math.cos(K4.to("rad").value)
-            + 1.591274 * math.cos(K5.to("rad").value)
-        ) * u.deg
-
-        J1 = (129.071773 + 19140.0328244 * T) * u.deg
-        J2 = (36.352167 + 38281.0473591 * T) * u.deg
-        J3 = (56.668646 + 57420.9295360 * T) * u.deg
-        J4 = (67.364003 + 76560.2552215 * T) * u.deg
-        J5 = (104.792680 + 95700.4387578 * T) * u.deg
-        J6 = (95.391654 + 0.5042615 * T) * u.deg
-
-        W = (
-            176.049863
-            + 350.891982443297 * d
-            + 0.000145 * math.sin(J1.to("rad").value)
-            + 0.000157 * math.sin(J2.to("rad").value)
-            + 0.000040 * math.sin(J3.to("rad").value)
-            + 0.000001 * math.sin(J4.to("rad").value)
-            + 0.000001 * math.sin(J5.to("rad").value)
-            + 0.584542 * math.sin(J6.to("rad").value)
-        ) * u.deg
-
-        return ra, dec, W
-
-
-class _Jupiter(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        Ja = (99.360714 + 4850.4046 * T) * u.deg
-        Jb = (175.895369 + 1191.9605 * T) * u.deg
-        Jc = (300.323162 + 262.5475 * T) * u.deg
-        Jd = (114.012305 + 6070.2476 * T) * u.deg
-        Je = (49.511251 + 64.3000 * T) * u.deg
-
-        ra = (
-            268.056595
-            - 0.006499 * T
-            + 0.000117 * math.sin(Ja.to("rad").value)
-            + 0.000938 * math.sin(Jb.to("rad").value)
-            + 0.001432 * math.sin(Jc.to("rad").value)
-            + 0.000030 * math.sin(Jd.to("rad").value)
-            + 0.002150 * math.sin(Je.to("rad").value)
-        ) * u.deg
-        dec = (
-            64.495303
-            + 0.002413 * T
-            + 0.000050 * math.cos(Ja.to("rad").value)
-            + 0.000404 * math.cos(Jb.to("rad").value)
-            + 0.000617 * math.cos(Jc.to("rad").value)
-            - 0.000013 * math.cos(Jd.to("rad").value)
-            + 0.000926 * math.cos(Je.to("rad").value)
-        ) * u.deg
-        W = (284.95 + 870.536 * d) * u.deg
-
-        return ra, dec, W
-
-
-class _Saturn(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        ra = (40.589 - 0.036 * T) * u.deg
-        dec = (83.537 - 0.004 * T) * u.deg
-        W = (38.90 + 810.7939024 * d) * u.deg
-
-        return ra, dec, W
-
-
-class _Uranus(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        ra = 257.311 * u.deg
-        dec = -15.175 * u.deg
-        W = (203.81 - 501.1600928 * d) * u.deg
-
-        return ra, dec, W
-
-
-class _Neptune(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        N = (357.85 + 52.316 * T) * u.deg
-
-        ra = (299.36 + 0.70 * math.sin(N.to("rad").value)) * u.deg
-        dec = (43.46 - 0.51 * math.cos(N.to("rad").value)) * u.deg
-        W = (249.978 + 541.1397757 * d - 0.48 * math.sin(N.to("rad").value)) * u.deg
-
-        return ra, dec, W
-
-
-class _Pluto(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        ra = 132.993 * u.deg
-        dec = -6.163 * u.deg
-        W = (302.695 + 56.3625225 * d) * u.deg
-
-        return ra, dec, W
-
-
-Mercury = _Mercury(
+Mercury = SolarSystemBody(
     parent=Sun,
     k=constants.GM_mercury,
     name="Mercury",
@@ -344,7 +141,7 @@ Mercury = _Mercury(
     rotational_period=constants.rotational_period_mercury,
 )
 
-Venus = _Venus(
+Venus = SolarSystemBody(
     parent=Sun,
     k=constants.GM_venus,
     name="Venus",
@@ -356,7 +153,7 @@ Venus = _Venus(
     J2=_q(constants.J2_venus),
     J3=_q(constants.J3_venus),
 )
-Earth = _Earth(
+Earth = SolarSystemBody(
     parent=Sun,
     k=constants.GM_earth,
     name="Earth",
@@ -369,7 +166,7 @@ Earth = _Earth(
     J2=_q(constants.J2_earth),
     J3=_q(constants.J3_earth),
 )
-Mars = _Mars(
+Mars = SolarSystemBody(
     parent=Sun,
     k=constants.GM_mars,
     name="Mars",
@@ -381,7 +178,7 @@ Mars = _Mars(
     J2=_q(constants.J2_mars),
     J3=_q(constants.J3_mars),
 )
-Jupiter = _Jupiter(
+Jupiter = SolarSystemBody(
     parent=Sun,
     k=constants.GM_jupiter,
     name="Jupiter",
@@ -392,7 +189,7 @@ Jupiter = _Jupiter(
     rotational_period=constants.rotational_period_jupiter,
     mass=_q(constants.M_jupiter),
 )
-Saturn = _Saturn(
+Saturn = SolarSystemBody(
     parent=Sun,
     k=constants.GM_saturn,
     name="Saturn",
@@ -402,7 +199,7 @@ Saturn = _Saturn(
     R_polar=constants.R_polar_saturn,
     rotational_period=constants.rotational_period_saturn,
 )
-Uranus = _Uranus(
+Uranus = SolarSystemBody(
     parent=Sun,
     k=constants.GM_uranus,
     name="Uranus",
@@ -412,7 +209,7 @@ Uranus = _Uranus(
     R_polar=constants.R_polar_uranus,
     rotational_period=constants.rotational_period_uranus,
 )
-Neptune = _Neptune(
+Neptune = SolarSystemBody(
     parent=Sun,
     k=constants.GM_neptune,
     name="Neptune",
@@ -423,7 +220,7 @@ Neptune = _Neptune(
     rotational_period=constants.rotational_period_neptune,
 )
 
-Pluto = _Pluto(
+Pluto = SolarSystemBody(
     parent=Sun,
     k=constants.GM_pluto,
     name="Pluto",
@@ -435,69 +232,7 @@ Pluto = _Pluto(
 )
 
 
-class _Moon(SolarSystemBody):
-    @staticmethod
-    def _rot_elements_at_epoch(T, d):
-        E1 = (125.045 - 0.0529921 * d) * u.deg
-        E2 = (250.089 - 0.1059842 * d) * u.deg
-        E3 = (260.008 + 13.0120009 * d) * u.deg
-        E4 = (176.625 + 13.3407154 * d) * u.deg
-        E5 = (357.529 + 0.9856003 * d) * u.deg
-        E6 = (311.589 + 26.4057084 * d) * u.deg
-        E7 = (134.963 + 13.0649930 * d) * u.deg
-        E8 = (276.617 + 0.3287146 * d) * u.deg
-        E9 = (34.226 + 1.7484877 * d) * u.deg
-        E10 = (15.134 - 0.1589763 * d) * u.deg
-        E11 = (119.743 + 0.0036096 * d) * u.deg
-        E12 = (239.961 + 0.1643573 * d) * u.deg
-        E13 = (25.053 + 12.9590088 * d) * u.deg
-
-        ra = (
-            269.9949
-            + 0.0031 * T
-            - 3.8787 * math.sin(E1.to("rad").value)
-            - 0.1204 * math.sin(E2.to("rad").value)
-            + 0.0700 * math.sin(E3.to("rad").value)
-            - 0.0172 * math.sin(E4.to("rad").value)
-            + 0.0072 * math.sin(E6.to("rad").value)
-            - 0.0052 * math.sin(E10.to("rad").value)
-            + 0.0043 * math.sin(E13.to("rad").value)
-        ) * u.deg
-        dec = (
-            66.5392
-            + 0.0130 * T
-            + 1.5419 * math.cos(E1.to("rad").value)
-            + 0.0239 * math.cos(E2.to("rad").value)
-            - 0.0278 * math.cos(E3.to("rad").value)
-            + 0.0068 * math.cos(E4.to("rad").value)
-            - 0.0029 * math.cos(E6.to("rad").value)
-            + 0.0009 * math.cos(E7.to("rad").value)
-            + 0.0008 * math.cos(E10.to("rad").value)
-            - 0.0009 * math.cos(E13.to("rad").value)
-        ) * u.deg
-        W = (
-            38.321
-            + 13.17635815 * d
-            - 1.4e-12 * d ** 2
-            + 3.5610 * math.sin(E1.to("rad").value)
-            + 0.1208 * math.sin(E2.to("rad").value)
-            - 0.0642 * math.sin(E3.to("rad").value)
-            + 0.0158 * math.sin(E4.to("rad").value)
-            + 0.0252 * math.sin(E5.to("rad").value)
-            - 0.0066 * math.sin(E6.to("rad").value)
-            - 0.0047 * math.sin(E7.to("rad").value)
-            - 0.0046 * math.sin(E8.to("rad").value)
-            + 0.0028 * math.sin(E9.to("rad").value)
-            + 0.0052 * math.sin(E10.to("rad").value)
-            + 0.0040 * math.sin(E11.to("rad").value)
-            + 0.0019 * math.sin(E12.to("rad").value)
-            - 0.0044 * math.sin(E13.to("rad").value)
-        ) * u.deg
-
-        return ra, dec, W
-
-
-Moon = _Moon(
+Moon = SolarSystemBody(
     parent=Earth,
     k=constants.GM_moon,
     name="Moon",
